requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
requests
loguru
typer
flask
pydantic-settings
bs4
gTTS
SpeechRecognition
pythainlp
pyaudio
python-crfsuite
resemblyzer
numpy
sounddevice
python-multipart
langdetect
google-cloud-texttospeech
pygame
mutagen
tuya-connector-python
pillow
python-dateutil
resampy
slugify
dateparser
langdetect
deep_translator
dateparser
slugify
feedparser